Smoke tests: fix live check url depending on deployment context
